pyproject.toml
[tool.poetry]
name = "kube-schedule-scaler"
version = "0.1.0"
description = ""
authors = ["Håvard Lindset <havard@hoopla.no>"]

[tool.poetry.dependencies]
python = "^3.8"
croniter = "^0.3.34"
kubernetes = "^11.0.0"
python-dateutil = "^2.8.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
